jwt 검증 voteDelete service

+++ src/main/java/com/example/manymanyUsers/vote/service/VoteService.java
@@ -97,7 +97,12 @@ public class VoteService {
 
     }
 
-    public void deleteVote(Long voteId) {
+    public void deleteVote(Long voteId, Long userId) throws NotFoundException {
+
+        Optional<User> find = userRepository.findById(userId);
+        if(find.isEmpty()) {
+            throw new NotFoundException("해당 아이디를 가진 유저가 없습니다. 아이디 값을 다시 한번 확인하세요.");
+        }
 
         voteRepository.deleteById(voteId);
